requirements.txt
fastapi~=0.112.3
httpx~=0.27.2 
pydantic~=2.8.2 
pydantic-settings~=2.4.0 
sqlalchemy-utils~=0.41.2 
sqlalchemy~=2.0.34 
starlette~=0.38.4 
uvicorn~=0.30.6
pyfirmata~=1.1.0
pyserial~=3.5
requests~=2.32.3
python-dotenv~=1.0.1
geopy~=2.4.1
pywhatkit~=5.4
pymysql~=1.0.0
cryptography~=40.0.0